backend/pyproject.toml
[project]
name = "duel-master-backend"
version = "1.0.0"
description = ""
authors = [{ name = "RogelioKG", email = "rogeliokg0813@gmail.com" }]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.0",
    "flask-cors>=5.0.0",
    "requests>=2.32.3",
    "azure-cognitiveservices-vision-computervision>=0.9.1",
    "transformers>=4.47.1",
    "torch",
    "dotenv>=0.9.9",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-cov>=6.0.0",
    "pytest-mock>=3.14.0",
    "python-docx>=1.1.2",
    "ruff>=0.9.10",
]

[tool.uv.sources]
torch = [{ index = "pytorch-cu124" }]

[[tool.uv.index]]
name = "pypi"
url = "https://pypi.org/simple/"

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"

[tool.ruff]
indent-width = 4
line-length = 100

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
    "UP", # pyupgrade
]

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
pythonpath = "src"
testpaths = "tests"

[tool.coverage.run]
data_file = "tests/.coverage"
omit = [
    "src/constanst.py",
    "src/shared_types.py",
    "src/utils/log.py"
]
